refactor(learning-sessions): split 515-LOC monolith into focused modules (Related to #1955)
Extract learning_sessions.py into three focused modules:
- sessions_bootstrap.py  — POST create (get-or-create, slug validation), list, reading-history
- sessions_progress.py   — GET detail, GET status, PATCH progress + CharacterError persistence
- sessions_complete.py   — GET /sessions/{id}/report alias

learning_sessions.py is now a thin facade that composes the three routers.
All 7 URL paths unchanged; alembic heads stays at 1 (no DB changes).

TDD: 20/20 characterization tests GREEN on monolith and after split.
Includes fix for conftest plain-string ::jsonb server_default not being
patched (learning_sessions.full_reading_attempts nullable=False).
